Move the YAML indent constant to a leaf module
