+++ src/util/yaml-section-lexer.ts
@@ -5,7 +5,7 @@
  * constant. (Scalar value parsing lives in yaml-scalar.ts.)
  */
 
-import { ESPHOME_YAML_INDENT } from "./esphome-yaml-lang.js";
+import { ESPHOME_YAML_INDENT } from "./esphome-yaml-indent.js";
 
 /** Non-leading key characters: anything but whitespace, ``:``, ``#``. */
 const KEY_BODY = "[^\\s:#]*";
