+++ test/util/config-entry-yaml-scan.test.ts
@@ -8,6 +8,7 @@ import {
   findReferenceCandidates,
   findUsedPins,
   isCertainlyDanglingId,
+  sectionEndLine,
   yamlHasExternalIdSources,
   yamlHasMergedSources,
 } from "../../src/util/config-entry-yaml-scan.js";
@@ -76,6 +77,21 @@ describe("findUsedPins", () => {
     const map = findUsedPins(yaml);
     expect(map.get(4)).toBe("switch");
     expect(map.get(5)).toBe("binary_sensor");
+  });
+
+  it("bounds sectionEndLine at the next sibling, dot and underscore keys included", () => {
+    const config = [
+      "switch:", //          line 1
+      "  - platform: gpio",
+      ".anchors:", //        line 3
+      "  - &spare 7",
+      "_extras:", //         line 5
+      "  note: x",
+      "",
+    ].join("\n");
+    expect(sectionEndLine(config, 1)).toBe(2);
+    expect(sectionEndLine(config, 3)).toBe(4);
+    expect(sectionEndLine(config, undefined)).toBeUndefined();
   });
 
   it("skips pins inside a dot-prefixed ignored block", () => {
